move Filters in container
