+++ src/assets/icons/icons/index.ts
@@ -21,5 +21,5 @@ export { PenIcon } from './components/Pen';
 export { CreateNFT } from './components/CreateNFT';
 export { LightTheme } from './components/LightTheme';
 export { Promo } from './components/Promo';
-export { Filters } from './components/Filters';
+export { FiltersIcon } from './components/Filters';
 export * from './socials';
