+++ src/containers/Filters/Filters.mock.tsx
@@ -1,0 +1,7 @@
+import { FiltersProps } from './Filters';
+
+export const filtersPropsMocked: FiltersProps = {
+  filters: {},
+  isShowFilters: false,
+  onClose: () => {},
+};
