+++ src/containers/Filters/index.tsx
@@ -1,0 +1,2 @@
+export * from './Filters';
+export * from './styles.module.scss';
