+++ src/containers/index.tsx
@@ -3,5 +3,6 @@ export * from './MobileNavigation';
 export * from './Footer';
 export * from './NotificationModal';
 export * from './RouterManager';
+export * from './Filters';
 export * from './Layout/Layout';
 export * from './FallbackImage/FallbackImage';
